# final/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Homework API
    description: BFF
    version: 0.0.1
paths:
    /find-answer:
        post:
            tags:
                - Homework
            description: 获取答案
            operationId: Homework_FindSolution
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/homework.v1.FindSolutionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/homework.v1.FindSolutionReply'
    /records/{userId}:
        get:
            tags:
                - Homework
            description: 搜索记录列表
            operationId: Homework_GetRecords
            parameters:
                - name: userId
                  in: path
                  description: 用户ID
                  required: true
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/homework.v1.GetRecordsReply'
components:
    schemas:
        homework.v1.FindSolutionReply:
            type: object
            properties:
                solveSolution:
                    $ref: '#/components/schemas/homework.v1.SolveSolution'
                searchSolution:
                    type: string
        homework.v1.FindSolutionRequest:
            type: object
            properties:
                content:
                    type: string
                    description: 题目内容
                fileUrl:
                    type: string
                    description: 题目图片链接
                userId:
                    type: integer
                    description: 用户ID
                    format: uint64
        homework.v1.GetRecordsReply:
            type: object
            properties:
                records:
                    type: array
                    items:
                        $ref: '#/components/schemas/homework.v1.GetRecordsReply_Record'
        homework.v1.GetRecordsReply_Question:
            type: object
            properties:
                content:
                    type: string
                    description: 题目内容
                fileUrl:
                    type: string
                    description: 题目图片链接
        homework.v1.GetRecordsReply_Record:
            type: object
            properties:
                question:
                    $ref: '#/components/schemas/homework.v1.GetRecordsReply_Question'
                solution:
                    $ref: '#/components/schemas/homework.v1.GetRecordsReply_Solution'
        homework.v1.GetRecordsReply_Solution:
            type: object
            properties:
                searchSolution:
                    type: string
                    description: 搜索结果
                solveSolution:
                    type: string
                    description: 解题结果
        homework.v1.SolveSolution:
            type: object
            properties:
                input:
                    type: string
                    description: 输入题目
                solution:
                    type: string
                    description: 答案
                steps:
                    type: array
                    items:
                        $ref: '#/components/schemas/homework.v1.SolveSolution_Step'
                    description: 步骤列表
                knowledge:
                    type: integer
                    description: 知识点
                    format: int32
        homework.v1.SolveSolution_Step:
            type: object
            properties:
                name:
                    type: string
                    description: 步骤名称
                output:
                    type: string
                    description: 步骤结果
            description: 步骤
tags:
    - name: Homework